sci-biology/micro-manager: version bump to 1.4.20 and update live ebuild
Re-pointed the live ebuild to the github mirror provided by mdcurtis, mainly
to avoid the requirement of subversion being installed, however it is also nice to
match where we obtain the source tarballs from.

Got the build system to accept system-installed jars for dependencies in the java code,
with a minmalist amount of patching.  Also created a launcher via java-pkg_dolauncher so
that the .desktop file actually points to something.

The ClojureEditor plugin is currently disabled as it requires clooj, and the ebuild
for that package and its dependencies is not yet ready.  Note in the future, some sort
of use flag structure will likely be added to enable or disable specific plugins.