sci-biology/micro-manager: fixed launchers and imagej integration
Prior attempts at the launcher in 1.4.20 and 1.4.21 would allow micro-manager to run
in a standalone mode but it was crippled, and there did not seem to be any way to get
it to work properly within ImageJ as a plugin the way it was designed to.  Turns out
all that was needed was to set the current working directory to the imagej libdir
prior to launch.

The main micro-manager launcher (and .desktop entry) now launches micro-manager
as an ImageJ plugin, and a second launcher 'micro-manager-standalone' is available
and to be run at the commandline.